Implement custom validation exception for CRUD operation of LeaveRequst, LeaveType and LeaveAllocation domain modeles

+++ HR.LeaveManagement.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
 using HR.LeaveManagement.Application.Persistence.Contracts;
 using MediatR;
@@ -21,6 +22,8 @@ namespace HR.LeaveManagement.Application.Features.LeaveRequests.Handlers.Command
         public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
         {
             var leaveRequest = await _leaveRequestRepository.GetAsync(request.Id);
+            if (leaveRequest is null)
+                throw new NotFoundException(nameof(Domain.LeaveRequest), request.Id);
 
             await _leaveRequestRepository.DeleteAsync(leaveRequest);
